[Polyam] Widen the clickable area for statuses in grouped notifications
Port a8330be93ee825948a17c810af33e28bfa2c56c8 to polyam-glitch
